modify migration table

+++ database/migrations/2023_10_20_160120_create_physical_examinations_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->integer('bot_bp');
             $table->integer('pulse');
             $table->integer('respiratory_rate');
-            $table->integer('bmi');
+            $table->float('bmi');
 
             //Secondary
             $table->string('pe_general_appearance')->default('No');
